Make JavaApp > BACNET_MASTER working
