+++ src/system/apps/base/frontend_app.py
@@ -1,4 +1,3 @@
-import os
 from abc import ABC
 
 from src.system.apps.base.installable_app import InstallableApp
@@ -18,12 +17,6 @@ class FrontendApp(InstallableApp, ABC):
     def select_asset(self, row: any):
         return row.get('zipball_url')
 
-    def after_download(self, download_name: str):
-        # they are already wrapped on folder
-        download_dir: str = self.get_download_dir()
-        extracted_dir = os.path.join(download_dir, download_name)
-        os.rename(extracted_dir, self.get_downloaded_dir())
-
     def install(self) -> bool:
         # TODO: remove user and upgrade parameters in future
         install_cmd: str = "sudo bash script.bash start -service_name={} -u={} -dir={} -data_dir={} -p={}".format(
@@ -32,8 +25,4 @@ class FrontendApp(InstallableApp, ABC):
 
     def uninstall(self) -> bool:
         uninstall_cmd: str = "sudo bash script.bash delete"
-        return execute_command(uninstall_cmd, self.get_cwd())
-
-    def restart(self) -> bool:
-        uninstall_cmd: str = f"sudo systemctl restart {self.service_file_name}"
         return execute_command(uninstall_cmd, self.get_cwd())
