+++ src/system/apps/base/java_app.py
@@ -1,10 +1,10 @@
 from abc import ABC, abstractmethod
 
-from src.system.apps.base.installable_app import InstallableApp
+from src.system.apps.base.systemd_app import SystemdApp
 from src.system.apps.enums.types import Types
 
 
-class JavaApp(InstallableApp, ABC):
+class JavaApp(SystemdApp, ABC):
     @property
     def app_type(self):
         return Types.JAVA_APP.value
@@ -18,15 +18,3 @@ class JavaApp(InstallableApp, ABC):
     @abstractmethod
     def name_contains(self) -> str:
         raise NotImplementedError("name_contains needs to be overridden")
-
-    def after_download(self, download_name: str):
-        pass
-
-    def install(self):
-        pass
-
-    def uninstall(self):
-        pass
-
-    def restart(self):
-        pass
