+++ src/system/apps/java/bacnet_master_app.py
@@ -1,4 +1,11 @@
+import os
+
+from flask import current_app
+from werkzeug.local import LocalProxy
+
 from src.system.apps.base.java_app import JavaApp
+
+logger = LocalProxy(lambda: current_app.logger)
 
 
 class BACnetMasterApp(JavaApp):
@@ -22,3 +29,13 @@ class BACnetMasterApp(JavaApp):
     @property
     def name_contains(self) -> str:
         return 'bacnet'
+
+    def create_service(self):
+        lines = []
+        with open(os.path.join(self.get_wd(), 'conf/nubeio-bacnet.service')) as systemd_file:
+            wd: str = self.get_wd()
+            for line in systemd_file.readlines():
+                if '/app/bacnet' in line:
+                    line = line.replace('/app/bacnet', wd)
+                lines.append(line)
+        return lines
